✨ Introduced dynamic post generation actions
Specific actions need to be taken when a project is generated (for example, committing the generated files and pushing them to the origin). However, the actions themselves are dynamic based on the options selected prior to generation (for example, we only need to change git execute permissions on scripts if git is the selected repository tool and the generation is happening on Windows (the permissions can be programmatically added on Linux and MacOS)). 

This change handles those scenarios through the `post_generation_actions.html` file, and postprocess actions (through `__postprocess.py` files) can customize what appears on that page.