🎨 [refactor] Refactor to better support changes
The current configuration supports the creation of different types of projects with different behavior:

- Generate documentation (yes/no)
- Repository tool (git, None)
- Hosting tool (GitHub, None)
- Project Type (Python Execution Environment, Python Package, None)

Some of these options impact files created by other options (examples: 1) a GitHub hosting tool impacts the way SECURITY.md is generated within the documentation 2) a Python Package project type impacts the way README.md is generated within the documentation 3) Selecting CI within the Python Package project type impacts the way DEVELOPMENT.md is generated within the documentation).

Currently, the different options have code to write to template regions defined by other options (for example, the GitHub hosting code writes into a template region defined by the SECURITY.md documentation file). The hope was that this would help to keep all of the GitHub-related logic in the same place. Unfortunately, this means that the code to fully construct SECURITY.md is spread throughout the code base.

In this refactor, I will change the logic so that the knowledge to fully construct files is located near the file itself. Hopefully, this change results in PRs that are simpler to review and logic that is easier to maintain long term.