🐛 [-bug] Removed .git directory validation due to copier update
Prior to this change, the code would check for a .git directory in the generated output if git was selected as the repository tool. Unfortunately, this check conflicted with the way that `copier update` determined the deltas between the new template and template previously used to generate the content (those temporary directories did not have a .git directory and failed the check).

This change removes the check.