🐛 [-bug] Ensure that context steps aren't invoked in copier update
During update, copier will fully generate different template versions to local directories and then create a diff for all changes between those generated directories. These diffs are then applied to the actual directory. 

We don't want the dynamic context steps to be invoked when generating the temporary directories, as the steps will generate different values in each, which will result in a diff applied to the actual directory that contains the dynamic values.

Detect when we are in this scenario and ensure the dynamic values are populated with the same placeholder values so that they don't result in a diff.