🐛 [-bug] Addressed issue with running minisign within a docker container
Previously, the context scripts would attempt to use docker to generate a minisign key at generation time. However, this didn't work when copier was run from within a docker container.

With this change, the code will attempt to execute a local version of minisign before reverting to using the minisign docker image.

The change also includes updates to the copier-projectscaffolding docker image to include minisign.