🐛 [-bug] Hopefully addressing update issue
The validators for nested conditional template variables were failing during update scenarios when the variables were never populated during the initial copy.

It looks like this is happening because the template variables were default initialized to an empty string. This didn't cause problems during `copier copy` but breaks things during `copier update`.

I have run a similar change in a test template and verified that this change works around the problem, but I need to code to be added to this template before I can validate that it works here.