🐛 [-bug] Fix markdown rendering issue
With the previous change, comments delimiting sections dynamically populated are preserved to support better git diffs. However, this caused a regression in how markdown tables are rendered, as the HTML comments were embedded within the table, breaking GitHub's ability to render the markdown table. The fix is to move the delimiting comment outside of the table definition.